+++ src/nl/hannahsten/texifyidea/run/bibtex/BibtexCommandLineState.kt
@@ -7,6 +7,9 @@ import com.intellij.execution.process.KillableProcessHandler
 import com.intellij.execution.process.ProcessHandler
 import com.intellij.execution.process.ProcessTerminatedListener
 import com.intellij.execution.runners.ExecutionEnvironment
+import com.intellij.util.execution.ParametersListUtil
+import nl.hannahsten.texifyidea.run.compiler.LatexCompiler.Companion.toWslPath
+import nl.hannahsten.texifyidea.run.latex.LatexDistributionType
 
 /**
  * @author Sten Wessel
@@ -19,7 +22,25 @@ open class BibtexCommandLineState(
     @Throws(ExecutionException::class)
     override fun startProcess(): ProcessHandler {
         val compiler = runConfig.compiler ?: throw ExecutionException("No valid compiler specified.")
-        val command: List<String> = compiler.getCommand(runConfig, environment.project) ?: throw ExecutionException("Compile command could not be created.")
+        val compilerCommand = compiler.getCommand(runConfig, environment.project)?.toMutableList() ?: throw ExecutionException("Compile command could not be created.")
+
+        // See LatexCompiler#getCommand
+        val command = if (runConfig.getLatexDistributionType() == LatexDistributionType.WSL_TEXLIVE) {
+            var wslCommand = GeneralCommandLine(compilerCommand).commandLineString
+
+            // Custom compiler arguments specified by the user
+            runConfig.compilerArguments?.let { arguments ->
+                ParametersListUtil.parse(arguments)
+                    .forEach { wslCommand += " $it" }
+            }
+
+            wslCommand += " ${runConfig.bibWorkingDir?.path?.toWslPath(runConfig.getLatexDistributionType())}"
+
+            mutableListOf("bash", "-ic", wslCommand)
+        }
+        else {
+            compilerCommand
+        }
 
         // The working directory is as specified by the user in the working directory.
         // The fallback (if null or empty) directory is the directory of the main file.