+++ src/nl/hannahsten/texifyidea/run/bibtex/BibtexRunConfiguration.kt
@@ -13,6 +13,8 @@ import com.intellij.openapi.vfs.LocalFileSystem
 import com.intellij.openapi.vfs.VirtualFile
 import nl.hannahsten.texifyidea.run.bibtex.logtab.BibtexLogTabComponent
 import nl.hannahsten.texifyidea.run.compiler.BibliographyCompiler
+import nl.hannahsten.texifyidea.run.latex.LatexDistributionType
+import nl.hannahsten.texifyidea.settings.sdk.LatexSdkUtil
 import org.jdom.Element
 
 /**
@@ -32,6 +34,7 @@ class BibtexRunConfiguration(
         private const val COMPILER_ARGUMENTS = "compiler-arguments"
         private const val MAIN_FILE = "main-file"
         private const val AUX_DIR = "aux-dir"
+        private const val LATEX_DISTRIBUTION = "latex-distribution"
     }
 
     var compiler: BibliographyCompiler? = null
@@ -45,6 +48,7 @@ class BibtexRunConfiguration(
 
     var mainFile: VirtualFile? = null
     var bibWorkingDir: VirtualFile? = null
+    internal var latexDistribution = LatexDistributionType.PROJECT_SDK
 
     override fun getConfigurationEditor(): SettingsEditor<out RunConfiguration> = BibtexSettingsEditor(project)
 
@@ -108,6 +112,15 @@ class BibtexRunConfiguration(
         else {
             null
         }
+
+        val distributionText = parent.getChildText(LATEX_DISTRIBUTION)
+        if (distributionText == null) {
+            // Make sure migration doesn't break the run configuration: keep the old behaviour
+            this.latexDistribution = LatexDistributionType.PROJECT_SDK
+        }
+        else {
+            this.latexDistribution = LatexDistributionType.valueOfIgnoreCase(distributionText)
+        }
     }
 
     override fun writeExternal(element: Element) {
@@ -122,6 +135,7 @@ class BibtexRunConfiguration(
         this.environmentVariables.writeExternal(parent)
         parent.addContent(Element(MAIN_FILE).apply { text = mainFile?.path ?: "" })
         parent.addContent(Element(AUX_DIR).apply { text = bibWorkingDir?.path ?: "" })
+        parent.addContent(Element(LATEX_DISTRIBUTION).also { it.text = latexDistribution.name })
     }
 
     override fun isGeneratedName() = name == suggestedName()
@@ -130,5 +144,19 @@ class BibtexRunConfiguration(
 
     fun setSuggestedName() {
         name = suggestedName()
+    }
+
+    // Similar to LatexRunConfiguration
+    fun setDefaultDistribution(distributionType: LatexDistributionType) {
+        latexDistribution = distributionType
+    }
+
+    fun getLatexDistributionType(): LatexDistributionType {
+        return if (latexDistribution != LatexDistributionType.PROJECT_SDK) {
+            latexDistribution
+        }
+        else {
+            LatexSdkUtil.getLatexDistributionType(project) ?: LatexDistributionType.TEXLIVE
+        }
     }
 }