+++ src/nl/hannahsten/texifyidea/run/compiler/LatexCompiler.kt
@@ -277,15 +277,6 @@ enum class LatexCompiler(private val displayName: String, val executableName: St
     ;
 
     /**
-     * Convert Windows paths to WSL paths.
-     */
-    private fun String.toPath(runConfig: LatexRunConfiguration): String =
-        if (runConfig.getLatexDistributionType() == LatexDistributionType.WSL_TEXLIVE) {
-            runCommand("wsl", "wslpath", "-a", this) ?: this
-        }
-        else this
-
-    /**
      * Get the execution command for the latex compiler.
      *
      * @param runConfig
@@ -315,7 +306,7 @@ enum class LatexCompiler(private val displayName: String, val executableName: St
             "/out"
         }
         else {
-            runConfig.outputPath.getAndCreatePath()?.path?.toPath(runConfig)
+            runConfig.outputPath.getAndCreatePath()?.path?.toWslPath(runConfig.getLatexDistributionType())
         }
 
         val auxilPath = if (runConfig.getLatexDistributionType() == LatexDistributionType.DOCKER_MIKTEX) {
@@ -325,7 +316,7 @@ enum class LatexCompiler(private val displayName: String, val executableName: St
             null
         }
         else {
-            runConfig.auxilPath.getAndCreatePath()?.path?.toPath(runConfig)
+            runConfig.auxilPath.getAndCreatePath()?.path?.toWslPath(runConfig.getLatexDistributionType())
         }
 
         val command = createCommand(
@@ -345,7 +336,7 @@ enum class LatexCompiler(private val displayName: String, val executableName: St
                     .forEach { wslCommand += " $it" }
             }
 
-            wslCommand += " ${mainFile.path.toPath(runConfig)}"
+            wslCommand += " ${mainFile.path.toWslPath(runConfig.getLatexDistributionType())}"
 
             return mutableListOf("bash", "-ic", wslCommand)
         }
@@ -481,6 +472,15 @@ enum class LatexCompiler(private val displayName: String, val executableName: St
                 it.executableName.equals(exe, true)
             } ?: PDFLATEX
         }
+
+        /**
+         * Convert Windows paths to WSL paths.
+         */
+        fun String.toWslPath(distributionType: LatexDistributionType): String =
+            if (distributionType == LatexDistributionType.WSL_TEXLIVE) {
+                runCommand("wsl", "wslpath", "-a", this) ?: this
+            }
+            else this
     }
 
     /**